decouple dom-walker from canvas re-render (#6211)
**Problem:**
Today dom-walker runs when the canvas re-renders. But instead we want to
be able to sometimes re-render the canvas and not run the dom-walker
(when we scroll the canvas, for example), and sometimes we want to
re-run the dom-walker but not re-render the canvas (when a
`RUN_DOM_WALKER` action is dispatched)

**Fix:**
- Decouple `shouldRerender` and `runDomWalker`.
- Revert `shouldRerender` to how it was a few months ago: if anything in
the dispatch result changes, re-render the canvas subtree. the Canvas
code has enough internal memoization to be resilient
- `shouldRunDOMWalker` instead of checking the full EditorState.canvas,
just checks a few important keys from it, most importantly, it does NOT
check the scroll offset.
- We want to be able to check a few keys of an object but ensure that we
provided an exhaustive list of what to check and what to deliberately
ignore – because otherwise we may silently fail to check important new
keys! To achieve this, I added a new `keysEqualityExhaustive` that uses
some typescript dark magic.

**Warning**
These changes were made in the parts of the dispatch which is not
testable by any of our test suites. (Something we should address in the
future), so please help in being vigilant and looking for potential
performance regressions this change might cause!