Do not allow internals of unfocused components to block selection (#6207)
**Problem:**
Context for the issue:
https://github.com/concrete-utopia/utopia/issues/5964
This is a followup to
https://github.com/concrete-utopia/utopia/pull/6158

The problem is described in detail in
https://paper.dropbox.com/doc/Selecting-elements-behind-the-internals-of-a-component--CUe0YryzTscIUHWtRh8AkhOkAg-suXRgxCUQhvLKLjQTv16F

**Summary:**
Components sometimes render internal elements on top of renderable
props. Typically our expectation is that we should be able to select the
children from props.children on the canvas. Even if a component has some
internals which cover those children, they should not block selection
when the component is unfocused. Why? Because those internals are not
selectable in an unfocused component. If you would like to edit the
internals, you need to focus the component anyway.

The disadvantage of this solution is that it is hard to focus on a
component in the canvas when its renderable props take up all of its
bounds. To edit the component you need to focus it in the navigator, or
using the context menu (and not with double click, which just selects a
renderable prop).

**Fix:**
Check in dom-lookup if the found element from elementsFromPoint is an
element from an unfocused component. If it is, treat it as if it was
locked (and continue the algorithm described in
https://github.com/concrete-utopia/utopia/pull/6158 )

**Manual Tests:**
I hereby swear that:

- [x] I opened a hydrogen project and it loaded
- [x] I could navigate to various routes in Preview mode

**Possible followup task:**
Move autofocusedPaths out of derived state and make it a memoized
function.

Fixes https://github.com/concrete-utopia/utopia/issues/5964