explicit canvas control ordering (#6219)
## Problem
it turns out that large controls can obscure small ones so that the
smaller control is still visible, but it cannot be grabbed because of
the large, invisible one.

Practical example: for CSS grids, we render a control that makes it
possible to drag a grid element, but in the case a CSS grid within a CSS
grid, the same grid outline control can overlay the grid-specific resize
controls, and make it impossible to grab them. However, if there was a
way to "force" some grid controls to the bottom of the control stack
this wouldn't be a problem

## Fix
Add a prop to `ControlWithProps` that makes it possible to force the
control to be near the top of the control stack, or near the bottom. If
the property is omitted, the control will be near the middle (between
the controls that have `bottom` specified and between the ones that have
`top` specified).