Fix CDN usage in webWorkerExtensionHost (#6233)
**Problem:**
VS Code now uses an iframe for running the web workers used for hosting
extensions. In staging we have VS Code use the CDN for loading
everything by providing the CDN as part of the base URL that is used via
its require function. That in turn meant that whilst VS Code was being
run in an iframe with the regular utopia domain, the extensions host
iframe was using cdn subdomain, which caused CORS issues.

**Fix:**
At the point where the iframe's URI is constructed we now strip the
`cdn.` from it's authority.
Included in this PR are some changes to local dev to use `cdn.localhost`
in all places where we would use the CDN subdomain, since it turns out
you don't need to do anything to make that work. ~~Also I was in the
process of tidying up some code after feedback from the previous PR, so
I've included that rather than separating it into another PR.~~

**Manual Tests:**
I hereby swear that:

- [x] I opened a hydrogen project and it loaded
- [x] I could navigate to various routes in Preview mode