Check that a valid GITHUB_TOKEN is supplied when running start-full (#6248)
**Problem:**
The VS Code build can fail due to GitHub rate limiting if the user
hasn't supplied a `GITHUB_TOKEN`, or if they have but it has expired.
When this happens, the failure message is very obscure, leaving the user
unclear how to progress.

**Fix:**
Previously we checked for the existence of the token, but that alone
isn't enough, since the an expired token will also cause the same
unclear failure case. Instead, we now perform a request against the GH
API using the token, and check that a 200 response code is received. If
no token was provided, or if the token has expired, a 40X response will
be received, in which case the process bails early and prints a helpful
message:

```
$ GITHUB_TOKEN=dirtman start-full                    
pg_ctl: PID file "utopia-db/postmaster.pid" does not exist
Is server running?
no server running on /private/tmp/tmux-501/default
A valid GITHUB_TOKEN is required when running the full Utopia build.
You may be missing one, or it may have expired.
Please see the readme for instructions.
```