Simplified auto flow options (#6259)
## Problem
When `Row dense` is set on then `Auto flow` dropdown, the dropdown shows
`Dense` as the selected option.

## Cause
It turns out that `gridAutoFlow: 'row dense'` makes its way to the style
prop just fine. However, in the DOM `row-dense` is turned into `dense`
(so the end product is `grid-auto-flow: dense`). This is not a
Utopia-specific problem, it's reproducible in CodeSandbox as well:
https://codesandbox.io/p/sandbox/muddy-glitter-djhdvv?file=%2Fsrc%2FApp.tsx%3A15%2C38.
We suspect that `row` and `row dense` means the same thing, so the
browser omits `row` from `row dense` to keep things simple.

## Fix
In order the make the `Auto flow` dropdown non-confusing (even though
it's technically correct as it is as long as the style prop is
concerned), this PR removes the `row dense` option, and only keeps
`dense`


### Manual Tests:
I hereby swear that:

- [x] I opened a hydrogen project and it loaded
- [x] I could navigate to various routes in Preview mode