remove spyMetadata and domMetadata references (#6266)
**Problem:**
A few places in the codebase refer to EditorState.domMetadata and
EditorState.spyMetadata. This was never supposed to be happening, but it
was never urgent to fix this, because both the domMetadata and
spyMetadata behaved in a reliable way. But the upcoming dom-walker
refactor pull requests will remove these helper states, and unify
everything into EditorState.jsxMetadata.

**Fix:**
To make the big PR cleaner, I am now deleting uses of spyMetadata and
domMetadata, and move all code to only use jsxMetadata