fixing a handful of setElementsToRerenderCommand (#6267)
**Problem:**
In the upcoming dom-walker refactor PR the dom-sampler will only collect
metadata for `ElementsToRerender` and their children. In testing we
realized that on `master` some behavior was hinging on the dom-walker
also collecting metadata for siblings and parent.

**Fix:**
Go through all uses of `setElementsToRerenderCommand` and make sure the
parent and siblings are included in the relevant strategies.