Spike: dom sampler (#6240)
**Problem:**
The current dom-walker's code is hard to read and hard to maintain, and
as a direct consequence of this, it is slow, does a lot of unnecessary
things and makes changing the ElementInstanceMetadata type an undue
burden.

The original dom-walker implemented a tree walk of the DOM using
recursive functions. The recursive tree walk unto itself is hard to
maintain, because VSCode can't spot unused parameters that are just
passed around. It was also necessary because 4 years ago we didn't have
the `data-path` printed into the DOM for each element, so the dom-walker
itself had to assemble paths based on the `data-uid`s it found.
 
Then on top of this code we grafted various layers of optimizations: one
layer attempts to cache results for each element while keeping the
recursive tree walk. As part of this optimization the dom-walker heavily
relies on its own previous results which need to be passed in.

The other problem is that the dom-walker is only half of the story, the
other half is the SpyWrapper system, which uses a react
higher-order-component wrapping every user element rendered by the
Canvas. Originally we had to keep these two data structures apart
because the canvas render was asynchronous, and we couldn't trust to run
the dom-walker only when the spy metadata was ready. But this constraint
has not been true for over two years now!

These two data sources are used to produce two data structures, which
are individually memoized, and then the two memoized data structures are
unified into a single structure in a costly step called
`reconstructJSXMetadata`.
`reconstructJSXMetadata` itself is hard to read and hard to maintain, so
unsurprisingly it is also pretty wasteful performance-wise, and changing
it is a MASSIVE pain in the ass every single time. Anecdotally, almost
everyone spent at least a few weeks on adding new types of elements that
required slightly different handling in the reconstruct function.

All in all, for a long time now the original constraints that made the
dom-walker as-is no longer applied, but refactoring the dom-walker into
something more manageable was always deemed a task too daunting.

**Fix:**
This PR introduces the `dom-sampler`. It tries to fix all of the above
problems:
- it's no longer a tree walk! turns out, ahead of time we have an
exhaustive array of ElementPaths of elements we want to collect metadata
for, called the `validPaths`. And we already print an attribute called
`data-path` for each dom element containint their ElementPath. Also
turns out, using `querySelectorAll` to query all relevant elements that
match a certain data-path is surprisingly quick! This means all of the
recursive tree walking functions can be skipped! This also means that
most of the hard-to-maintain code is gone.
- No more `reconstructJSXMetadata`! the dom-sampler returns a final
JsxMetadata map, and the necessary merging of the spy and dom
information sources are done internally during the processing of each
element. There are a few important features that were implemented as
part of `reconstructJSXMetadata`, namely `fillMissingDataFromAncestors`.
This is now brought into the dom-sampler as a step of fixes that can be
extended in the future. **Note**: this means that the dom-sampler uses
the spyMetadata as one of its inputs! This is not an issue because we
_synchronously_ render the canvas before we run the dom-sampler. This
was one of the keys to deleting a lot of code.

**Notes:**
- There is a new `EditorStore.elementMetadata` which for now we only use
to feed the dom-sampler with the previous results as a cache source. BUT
in the future, ALL other metadata from EditorState should be deleted,
further simplifying the update flow and the `dispatch.tsx` code
- I am yet to make conclusive measurements about this, but it looks like
`getAttributesComingFromStyleSheets` is 1 - 1.5ms – this should not even
be in the metadata, because we only use this information in the
inspector and only for the selected element, and we only want to update
it once every few seconds, yet FOR EACH FRAME we spend 1ms calculating
it. (optimizations like this can now be much easier implemented because
the new code is easier to change and measure)
- As I was reimplementing some of the features I realized we should
delete the whole concept of `infinityCanvasRectangle`, it was a well
meaning attempt to make the canvas appear infinite for the sake of the
canvas controls, but in reality it makes a lot of frame handling
unnecessarily complicated, and the canvas does have a size, although it
can expand infinitely, but if an element on the canvas root sizes itself
`left: 0, right: 0, top: 0, bottom: 0` it very much will have a size on
the screen which is not infinite.

**Befores and afters:**

A full re-walk of all elements, uncached:
dom-walker 67ms: 
<img width="1022" alt="image"
src="https://github.com/user-attachments/assets/2cae1595-134e-43dd-8df7-dff9ad6b2455">

dom-sampler 32ms:
<img width="631" alt="image"
src="https://github.com/user-attachments/assets/95cfb5e5-8828-4a98-bef9-a78431dd7e4f">

The frame when a reparent happens (only re-measuring the potentially
affected elements):
dom-walker 26ms:
<img width="726" alt="image"
src="https://github.com/user-attachments/assets/6d5f42cc-0703-4b9b-a08e-d5db32c6f59e">

dom-sampler 11.5ms:
<img width="488" alt="image"
src="https://github.com/user-attachments/assets/0e0c9707-b909-4068-9fca-6642f924f3dc">

Moving or resizing a single element (only re-measuring the one moved /
resized element):
dom-walker 18ms:
<img width="929" alt="image"
src="https://github.com/user-attachments/assets/cab3d642-e6a8-4342-afbd-a8df4efb8b15">

 dom-sampler 6ms:
<img width="692" alt="image"
src="https://github.com/user-attachments/assets/18090335-dfd4-483d-8837-cad9ce845e7d">

---------

Co-authored-by: Balint Gabor <127662+gbalint@users.noreply.github.com>