Cache getComponentDescriptorForTarget (#6294)
**Problem:**
In deriveState we check the autofocusability of all elements. To do that
we need to call getComponentDescriptorForTarget for each of them, which
is quite slow.
The component descriptor only depends on projectContents and
propertyControlsInfo, so we can use the new `valueDependentCache` to
cache this function. We just need to extend it to support a custom
equality function for the value.
With this change the time to check autofocusability of all elements in
derivestate is reduced from 1-2ms to ~0.3ms when there is not project
content change (e.g. during scrolling).

**Manual Tests:**
I hereby swear that:

- [x] I opened a hydrogen project and it loaded
- [x] I could navigate to various routes in Preview mode