NumberInput: holding arrow up-down makes the updates transient (#6378)
**Problem:**
The `NumberInput` component which is used all over the inspector did not
properly handle when the arrow up/down keys were pressed and held.

**Fix:**
It _looks like_ this did not work in the last 4 years! Amazingly, all of
the proper handling _was_ wired in, but then instead of passing in a
true flag, the code passed in a false flag, which made every event
non-transient. Weird!

**Commit Details:**
- Just passing in `true` as `incrementBy`'s transient parameter