LocalSelectedViews 2: Electric Boogaloo (#6411)
**Problem:**

Selection change can be upwards of 100ms, in a large enough project it
may reach half a second. This is because almost all parts of the Editor
UI use EditorState.selectedViews as a main part of their state, which
means changing the selected elements triggers a re-render of all of the
editor. For historical reasons, we also use this as an opportunity to
re-render and re-measure the canvas, correcting all metadata that may be
stale because of partial updates during interactions.

**Fix:**
The idea is that instead of speeding up the actual rerender, we make the
selection outline change first in a fast frame, then run the expensive
editor UI update in the next frame. Turns out that _perceptually_ this
is the desired behavior: the user looks at where the mouse is, they see
a very responsive UI change, and after having clicked with the mouse, we
have a 200-300ms deadzone of user perception before that can click or
move the mouse again. We fill that space with the expensive state
update, by the time the user moves their mouse or clicks again, the
editor is ready again to handle events responsively.

**Commit details**
- Adding `temporarilyDisableStoreUpdates` to `editor.tsx`
- Adding 'canvas-fast-selection-hack' and
'resume-canvas-fast-selection-hack' to the dispatch function's
"priority" parameter, which toggles `temporarilyDisableStoreUpdates` to
true or false.
- `mouseHandler` in `select-mode-hooks.tsx` dispatches
'canvas-fast-selection-hack' and then TWO animatonFrames later, a
'resume-canvas-fast-selection-hack'.
- Some selectors and hooks, most notably
`useGetApplicableStrategyControls` get a separate selectedViews field,
instead of using EditorState.selectedViews (which may be stale). This
selectedViews parameter is fed from localSelectedViews which is updated
sooner than the zustand editor store is. this is what enables the canvas
controls to draw the changed controls immediately while the rest of the
editor UI is temporarily one frame behind.

**Notes:**
- The tests DO NOT use the `temporarilyDisableStoreUpdates` system,
because the test's fake mouse helper events we dispatch are all fired in
a single animation frame (unlike Chrome's real mouse events). Changing
all this felt like too invasive for little benefit.
- IF we find event listeners that fire out of order or with stale state,
we can create a fourth Zustand store, and delete the canvas controls
localSelectedViews / localHighlightedViews. this fourth zustand store
would be immediately updated, and select important parts of the UI could
subscribe to it to get instant updates of the selection change.