TransientActions: making elementsToRerender mandatory (#6428)
**Problem:**
TransientActions is the inspector's way of creating a continuous
interaction. This properly collates all undo events into a single undo
step representing the entire series of dispatched transient actions, and
it _optionally_ lets you set an array of ElementPaths that will make the
editor and the canvas switch to fast rerender mode. The problem is that
this elementsToRerender path is optional, which means some inspector
interactions will be slow!

**Fix:**
Make elementsToRerender mandatory and fill it in all places.