Fix reparenting into grid (#6475)
**Problem:**
Reparenting into grid containers is broken.

**Fix:**
To reparent into a grid container we need grid cell measurements. The
dom sampler can only measure the grid cells when the controls are
displayed. The controls are displayed on hover during the interaction,
so this is is possible, but recently strategies were optimized by always
providing elementsToRerender values, which did not include the new
parent. So the grid controls were displayed, but not measured.
In case of grid reparent we need to make sure that the dom sampler can
measure the new parent, so the intendedParentPath has to be in
elementsToRerender.

**Manual Tests:**
I hereby swear that:

- [x] I opened a hydrogen project and it loaded
- [x] I could navigate to various routes in Play mode